fix(config): harden SafeConfig, slider, and reorderable list
- SafeConfig: add configType overload for validateOrReset (was hardcoded to -client.toml), add getDouble(), make WARNED set thread-safe, broaden exception handling, log secondary IOException
- ConfigSlider: guard against min==max division by zero, clamp initial to [0,1]
- ReorderableListScreen: extract hit zone constants, make EntryData.enabled final
- ConfigTab: use getDouble in percentSlider to avoid float precision loss

+++ src/main/java/xyz/kwahson/core/config/ConfigTab.java
@@ -126,7 +126,7 @@ public class ConfigTab extends GridLayoutTab {
   public ConfigSlider percentSlider(String label, double min, double max,
                                     ModConfigSpec.DoubleValue configVal) {
     return ConfigSlider.ofPercent(COL_WIDTH, label, min, max,
-        SafeConfig.getFloat(configVal, (float) ((min + max) / 2)), val -> configVal.set(val));
+        SafeConfig.getDouble(configVal, (min + max) / 2), val -> configVal.set(val));
   }
 
   public Button button(String label, Consumer<Button> onPress) {
